+++ src/pages/mint.tsx
@@ -9,12 +9,12 @@ import * as IPFS from 'ipfs-core';
 import { generateNftToken } from 'api/raribleApi';
 import { CONTRACT_ID } from 'utils/constants';
 import { useWallet } from 'wallet/state';
-import { CreateNftMetadata } from 'api/mintStructure';
 import { Web3Ethereum } from '@rarible/web3-ethereum';
 import { createRaribleSdk } from '@rarible/protocol-ethereum-sdk';
 import { SellRequest } from '@rarible/protocol-ethereum-sdk/build/order/sell';
 import { toAddress, toBigNumber } from '@rarible/types';
 import { useRouter } from 'next/router';
+import { CreateNftMetadata } from 'api/raribleRequestTypes';
 
 const MintPage = () => {
   const form = useForm();
